+++ lib/src/code_generators/swagger_enums_generator.dart
@@ -35,7 +35,8 @@ abstract class SwaggerEnumsGenerator extends SwaggerGeneratorBase {
     final enumsFromRequestBodies = generateEnumsFromSchemaMap(requestBodies);
 
     if (definitions.isEmpty) {
-      return '$enumsFromRequests$enumsFromResponses$enumsFromRequestBodies';
+      return '''import 'package:json_annotation/json_annotation.dart';
+$enumsFromRequests$enumsFromResponses$enumsFromRequestBodies''';
     }
 
     final enumsFromClasses = definitions.keys
@@ -110,9 +111,9 @@ $enumsFromRequestBodies
       final isInteger =
           kIntegerTypes.contains(swaggerRequestParameter.schema?.type) ||
               kIntegerTypes.contains(swaggerRequestParameter.items?.type);
-      
-      
-      if (enumValues.isNotEmpty && swaggerRoot.components?.schemas.containsKey(key) != true) {
+
+      if (enumValues.isNotEmpty &&
+          swaggerRoot.components?.schemas.containsKey(key) != true) {
         final enumContent = generateEnumContent(
           getValidatedClassName(key),
           enumValues,