+++ lib/src/code_generators/swagger_models_generator.dart
@@ -39,6 +39,10 @@ abstract class SwaggerModelsGenerator extends SwaggerGeneratorBase {
     List<String> allEnumListNames,
     Map<String, SwaggerSchema> allClasses,
   ) {
+    if (options.overridenModels.contains(getValidatedClassName(className))) {
+      return '';
+    }
+
     if (schema.isEnum) {
       return '';
     }
@@ -542,6 +546,7 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
     required List<String> allEnumListNames,
     required String propertyName,
     required List<String> requiredProperties,
+    required Map<String, String> basicTypesMap,
   }) {
     final allOf = prop.allOf;
     String typeName;
@@ -557,6 +562,10 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
       }
 
       typeName = getValidatedClassName(className);
+    }
+
+    if (basicTypesMap.containsKey(typeName)) {
+      typeName = basicTypesMap[typeName]!;
     }
 
     final includeIfNullString = generateIncludeIfNullString();
@@ -634,7 +643,15 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
 
     typeName = nullable(typeName, className, requiredProperties, propertyKey, prop);
 
-    if (options.classesWithNullabeLists.contains(className) && typeName.startsWith('List<') && !typeName.endsWith('?')) {
+    final propertySchema = allClasses[prop.ref.getUnformattedRef()];
+
+    if (propertySchema?.isNullable == true) {
+      typeName = typeName.makeNullable();
+    }
+
+    if (options.classesWithNullabeLists.contains(className) &&
+        typeName.startsWith('List<') &&
+        !typeName.endsWith('?')) {
       typeName += '?';
     }
 
@@ -730,7 +747,7 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
             allClasses: allClasses,
             prop: items,
             propertyName: propertyName,
-          ).asList();
+          ).makeNullable().asList();
         }
       }
 
@@ -1088,7 +1105,7 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
     if (propertiesMap.isEmpty) {
       return '';
     }
-
+    
     final results = <String>[];
     final propertyNames = <String>[];
 
@@ -1106,7 +1123,6 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
       propertyName = getParameterName(propertyName, propertyNames);
 
       propertyNames.add(propertyName);
-
       if (prop.type.isNotEmpty) {
         results.add(generatePropertyContentByType(
           prop,
@@ -1130,6 +1146,7 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
             allEnumNames: allEnumNames,
             propertyKey: propertyKey,
             propertyName: propertyName,
+            basicTypesMap: basicTypesMap,
             requiredProperties: requiredProperties,
           ),
         );
@@ -1169,7 +1186,7 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
     return results.join('\n');
   }
 
-  static Map<String, String> generateBasicTypesMapFromSchemas(SwaggerRoot root) {
+  Map<String, String> generateBasicTypesMapFromSchemas(SwaggerRoot root) {
     final result = <String, String>{};
 
     final components = root.components;
@@ -1197,7 +1214,16 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
         if (result[ref.getUnformattedRef()] != null) {
           result[key] = result[ref.getUnformattedRef()]!.asList();
         } else if (ref.isNotEmpty) {
-          result[key] = ref.getRef().asList();
+          var typeName = ref.getUnformattedRef();
+          final schema = allClasses[typeName];
+
+          if (kBasicTypes.contains(schema?.type)) {
+            typeName = _mapBasicTypeToDartType(schema!.type, value.format);
+          } else {
+            typeName = getValidatedClassName(typeName);
+          }
+
+          result[key] = typeName.asList();
         }
       }
     });
@@ -1365,6 +1391,7 @@ List<enums.$neededName>? ${neededName.camelCase}NullableListFromJson(
     Map<String, SwaggerSchema> allClasses,
   ) {
     final properties = getModelProperties(schema, schemas, allClasses);
+
     final requiredProperties = _getRequired(schema, schemas);
 
     final generatedConstructorProperties = generateConstructorPropertiesContent(
@@ -1578,7 +1605,9 @@ $allHashComponents;
 
     final newModelMap = allOf.firstWhereOrNull((m) => m.properties.isNotEmpty);
 
-    final currentProperties = newModelMap?.properties ?? {};
+    final currentProperties = schema.properties;
+
+    currentProperties.addAll(newModelMap?.properties ?? {});
 
     final refs = allOf.where((element) => element.ref.isNotEmpty).toList();
     for (var allOf in refs) {
@@ -1665,6 +1694,12 @@ $allHashComponents;
           final allOfModel = schemas[ref.getUnformattedRef()];
 
           final allOfModelProperties = allOfModel?.properties ?? {};
+
+          if (allOfModel?.allOf.isNotEmpty == true) {
+            for (var element in allOfModel?.allOf ?? <SwaggerSchema>[]) {
+              properties.addAll(element.properties);
+            }
+          }
 
           properties.addAll(allOfModelProperties);
         }